+++ data/web/Javascript.js
@@ -6,7 +6,9 @@
 
 const gpio_disabled = [];
 
-const gpio = [  {port: 2 , name:'D2'},
+const gpio = [  {port: 1,  name:'D1/TX0'},
+                {port: 2 , name:'D2'},
+                {port: 3,  name:'D3/RX0'},
                 {port: 4 , name:'D4'},
                 {port: 5 , name:'D5'},
                 {port: 6 , name:'D6'},