+++ src/baseconfig.h
@@ -23,6 +23,9 @@ class BaseConfig {
     const bool&     GetUseETH()        const { return useETH; }
     const String&   GetLANBoard()      const {return LANBoard;}
     const uint8_t&  GetDebugLevel()    const {return debuglevel;}
+    const uint8_t&  GetSerialRx()      const {return serial_rx;}
+    const uint8_t&  GetSerialTx()      const {return serial_tx;}
+
     const String    GetReleaseName(); 
   private:
     String    mqtt_server;
@@ -35,6 +38,8 @@ class BaseConfig {
     bool      useETH;  // otherwise use WIFI
     String    LANBoard;
     uint8_t   debuglevel;
+    uint8_t   serial_rx;
+    uint8_t   serial_tx;
 
 };