+++ ChangeLog.md
@@ -1,6 +1,8 @@
 Release 3.2.3:
   - add WebSerial as remote serial output as its own variant (#74)
   - add: configurable serial output pins
+  - replace deprecated arduinojson library "containsKey"
+  - use serveStatic function instead of custom handling
 
 Release 3.2.2:
   - add GoodWe Support, by @TigerGrey (#58)